adk-agent/pyproject.toml
[project]
name = "intent-to-agent"
version = "0.0.1"
description = "natural language intents into functional agents"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "google-adk==1.3.0",
    "python-dotenv==1.1.0",
    "httpx==0.28.1",
    "google-auth==2.40.3",
    "locust==2.37.10",
    "litellm==1.76.3",
]

[dependency-groups]
dev = [
    "pytest==8.4.0",
    "ruff==0.11.13",
]